test(preflight): split test_checks_diagnostics.py, empty allowlist
Split 302-line test file into 4 smaller modules (<85 lines each):
- test_checks_diagnostics_basic.py: basic checks + section header
- test_checks_diagnostics_api_health.py: API health checks
- test_checks_diagnostics_broker.py: broker readiness checks
- test_checks_diagnostics_market.py: market data + warn-only mode

Created conftest.py with shared fixtures:
- mock_healthy_client, checker, force_remote_env, warn_only_env

Removed last entry from check_test_hygiene.py ALLOWLIST.
