+++ var/agents/testing/index.json
@@ -8,7 +8,7 @@
   },
   "summary": {
     "total_tests": 6334,
-    "total_files": 977,
+    "total_files": 980,
     "markers_used": 15
   },
   "quick_commands": {
